Move dropdown classes and functions to separate files
