+++ src/constants.ts
@@ -1,0 +1,28 @@
+
+// Define dropdowns name that will be replaced by Field Finder UI
+export const dropdownsIncluded = [
+    "rffield",
+    "filterfilter",
+    "sort1",
+    "sort2",
+    "sort3",
+    "field",
+    "dffield",
+    "fffilter"
+];
+
+export const machinesThatSupportMultiSelect = [
+    "returnfields",
+    "filterfields",
+    "detailfields"
+];
+
+// Map of machine name to ID of field that contains field info
+export const relatedTableDataIds:any = {
+    "filters": "fffilter",
+    "returnfields": "rffield",
+    "summaryfilters": "filterfilter",
+    "detailfields": "dffield",
+    "filterfields": "fffilter",
+    "field": "field"
+};
